+++ hw_4_logs_and_test/src/main/java/ua/com/alevel/service/impl/PatientServiceImpl.java
@@ -1,6 +1,7 @@
 package ua.com.alevel.service.impl;
 
-import ua.com.alevel.dao.PatientDao;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import ua.com.alevel.dao.impl.PatientDaoImpl;
 import ua.com.alevel.entity.Patient;
 import ua.com.alevel.service.PatientService;
@@ -8,10 +9,15 @@ import ua.com.alevel.util.MyList;
 
 public class PatientServiceImpl implements PatientService {
 
+    private static final Logger LOGGER_INFO = LoggerFactory.getLogger("info");
+    private static final Logger LOGGER_WARN = LoggerFactory.getLogger("warn");
+    private static final Logger LOGGER_ERROR = LoggerFactory.getLogger("error");
     private final PatientDaoImpl patientDao = new PatientDaoImpl();
 
     public void create(Patient patient) {
+        LOGGER_INFO.info("patient start created");
         patientDao.create(patient);
+        LOGGER_INFO.info("patient finish created, id: " + patient.getId());
     }
 
     public void update(Patient patient) {
@@ -19,7 +25,25 @@ public class PatientServiceImpl implements PatientService {
     }
 
     public void delete(String id) {
-        patientDao.delete(id);
+        PatientService patientService = new PatientServiceImpl();
+        MyList<Patient> patients = patientService.findAll();
+        boolean canBeDeleted = true;
+        for (int i = 0; i < patients.getCountOfEntities(); i++) {
+            if (patients.getEntity(i) != null && patients.getEntity(i).getId().equals(id)) {
+                if (canBeDeleted) {
+                    System.out.println("You cannot delete a patient. Delete the declaration first:");
+                    canBeDeleted = false;
+                }
+                System.out.println(patients.getEntity(i).getId());
+            }
+        }
+        if (canBeDeleted) {
+            try {
+                patientDao.delete(id);
+            } catch (RuntimeException e) {
+                throw e;
+            }
+        }
     }
 
     public Patient findById(String id) {
@@ -37,5 +61,4 @@ public class PatientServiceImpl implements PatientService {
             throw exception;
         }
     }
-    public void showAllToConsole() { patientDao.showAllToConsole();}
 }
